chore: update package dependencies and versions
- Bump versions of several dependencies in package.json for improved stability and features.
- Update README files to reflect changes in usage and configuration.
- Refactor authentication logic to remove UserRepository and utilize TokenRepository for user data retrieval.
- Modify login route configuration to remove leading slash for consistency.
- Adjust API endpoints in constants to align with backend changes.
- Enhance error handling in HTTP interceptors to ensure proper redirection on unauthorized access.
- Remove deprecated UserRepository and related logic.
- Update notification service imports to use new provider method.
- Improve utility documentation and examples for better clarity.
