+++ apps/demo-app/src/app/constants/app.constants.ts
@@ -2,8 +2,8 @@ export const APP_CONSTANTS = {
   APP_NAME: 'Test App',
   VERSION: '1.0.0',
   API_ENDPOINTS: {
-    USERS: '/api/users',
-    PRODUCTS: '/api/products',
+    USERS: 'users',
+    PRODUCTS: 'products',
   },
   PAGINATION: {
     DEFAULT_PAGE_SIZE: 10,
