+++ package.json
@@ -29,9 +29,9 @@
   "dependencies": {
     "@acontplus/core": "^1.0.15",
     "@acontplus/ng-components": "^1.0.13",
-    "@acontplus/ng-config": "^1.0.3",
-    "@acontplus/ng-infrastructure": "^1.0.4",
-    "@acontplus/ng-notifications": "^1.0.8",
+    "@acontplus/ng-config": "^1.0.4",
+    "@acontplus/ng-infrastructure": "^1.0.5",
+    "@acontplus/ng-notifications": "^1.0.9",
     "@acontplus/utils": "^1.0.7",
     "@angular/animations": "20.3.2",
     "@angular/cdk": "20.2.5",
@@ -48,12 +48,12 @@
     "@jsverse/transloco": "^8.0.2",
     "bootstrap": "^5.3.8",
     "decimal.js": "^10.6.0",
-    "express": "^4.21.2",
+    "express": "^5.1.0",
     "jsonwebtoken": "9.0.2",
     "jwt-decode": "^4.0.0",
     "ngx-toastr": "^19.1.0",
     "rxjs": "^7.8.2",
-    "sweetalert2": "^11.23.0",
+    "sweetalert2": "^11.24.0",
     "tabulator-tables": "^6.3.1",
     "tslib": "^2.8.1",
     "uuid": "^13.0.0",
@@ -64,7 +64,7 @@
     "@angular-devkit/core": "20.3.3",
     "@angular-devkit/schematics": "20.3.3",
     "@angular/build": "20.3.3",
-    "@angular/cli": "20.3.3",
+    "@angular/cli": "~20.3.0",
     "@angular/compiler-cli": "20.3.2",
     "@jest/globals": "^30.2.0",
     "@nx/angular": "21.6.3",
@@ -78,22 +78,22 @@
     "@nx/vite": "21.6.3",
     "@nx/web": "21.6.3",
     "@nx/workspace": "21.6.3",
-    "@playwright/test": "^1.55.1",
+    "@playwright/test": "^1.56.0",
     "@schematics/angular": "^20.3.3",
     "@swc-node/register": "1.11.1",
     "@swc/core": "1.13.20",
     "@swc/helpers": "0.5.17",
-    "@types/express": "^4.17.21",
+    "@types/express": "^5.0.3",
     "@types/jest": "^30.0.0",
     "@types/jsonwebtoken": "9.0.10",
     "@types/node": "20.19.9",
     "@types/tabulator-tables": "^6.2.11",
-    "@typescript-eslint/utils": "^8.45.0",
+    "@typescript-eslint/utils": "^8.46.0",
     "@vitest/ui": "^3.2.4",
     "ajv": "^8.17.1",
     "angular-eslint": "20.3.0",
     "autoprefixer": "^10.4.21",
-    "eslint": "^9.36.0",
+    "eslint": "^9.37.0",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-playwright": "^2.2.2",
     "eslint-plugin-prettier": "^5.5.4",
@@ -101,7 +101,7 @@
     "jest": "^30.2.0",
     "jest-environment-jsdom": "^30.2.0",
     "jest-environment-node": "^30.2.0",
-    "jest-preset-angular": "^15.0.1",
+    "jest-preset-angular": "^15.0.2",
     "jest-util": "^30.2.0",
     "jiti": "2.4.2",
     "jsonc-eslint-parser": "^2.4.1",
@@ -114,10 +114,10 @@
     "rimraf": "^6.0.1",
     "ts-jest": "^29.4.4",
     "ts-node": "10.9.1",
-    "typescript": "^5.9.2",
-    "typescript-eslint": "^8.45.0",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.46.0",
     "verdaccio": "^6.2.0",
-    "vite": "^7.1.7",
+    "vite": "^7.1.9",
     "vite-plugin-dts": "4.5.4",
     "vitest": "^3.2.4"
   },
