+++ packages/ng-auth/package.json
@@ -1,15 +1,15 @@
 {
   "name": "@acontplus/ng-auth",
-  "version": "1.1.3",
+  "version": "1.1.4",
   "description": "Acontplus Angular Authentication Module",
   "dependencies": {
-    "@acontplus/ng-infrastructure": "^1.0.4",
-    "@acontplus/ng-config": "^1.0.3",
+    "@acontplus/ng-infrastructure": "^1.0.5",
+    "@acontplus/ng-config": "^1.0.4",
     "jwt-decode": "^4.0.0",
     "rxjs": "^7.8.1"
   },
   "peerDependencies": {
-    "@acontplus/core": "^1.0.13",
+    "@acontplus/core": "^1.0.15",
     "@angular/common": "^20.3.2",
     "@angular/core": "^20.3.2",
     "@angular/router": "^20.3.2",
