+++ packages/ng-auth/src/lib/application/use-cases/refresh-token.use-case.ts
@@ -1,7 +1,7 @@
 // src/lib/application/use-cases/refresh-token.use-case.ts
 import { Injectable, inject } from '@angular/core';
 import { Observable, throwError, catchError, tap } from 'rxjs';
-import { BaseUseCase, UserRepository } from '@acontplus/ng-infrastructure';
+import { BaseUseCase } from '@acontplus/ng-infrastructure';
 import { AuthRepository } from '../../domain/repositories/auth.repository';
 import { TokenRepository } from '../../repositories/token.repository';
 import { AuthTokens } from '@acontplus/core';
@@ -12,12 +12,11 @@ import { AuthStore } from '../../ui/stores/auth.store';
 })
 export class RefreshTokenUseCase extends BaseUseCase<void, AuthTokens> {
   private readonly authRepository = inject(AuthRepository);
-  private readonly userRepository = inject(UserRepository);
   private readonly tokenRepository = inject(TokenRepository);
   private readonly authStore = inject(AuthStore);
 
   execute(): Observable<AuthTokens> {
-    const userData = this.userRepository.getCurrentUser();
+    const userData = this.tokenRepository.getUserData();
     const refreshToken = this.tokenRepository.getRefreshToken();
 
     if (!userData?.email || !refreshToken || refreshToken.trim().length === 0) {
