+++ packages/ng-config/src/lib/constants/auth.constants.ts
@@ -1,5 +1,4 @@
 // src/lib/constants/auth.constants.ts
 export const AUTH_API = {
-  // Include trailing slash so repositories construct endpoints like '/account/login'
-  ACCOUNT: '/account/',
+  AUTH: 'auth/',
 } as const;
