+++ packages/ng-config/src/lib/environment.token.ts
@@ -8,10 +8,10 @@ import { Environment } from '@acontplus/core';
 export const ENVIRONMENT = new InjectionToken<Environment>('ENVIRONMENT', {
   factory: () => ({
     isProduction: false,
-    apiBaseUrl: 'http://localhost:3000/api',
+    apiBaseUrl: 'http://localhost:4200/api/',
     tokenKey: 'access_token',
     refreshTokenKey: 'refresh_token',
-    clientId: 'local-dev',
-    loginRoute: '/auth',
+    clientId: 'angular-app',
+    loginRoute: 'auth',
   }),
 });
