+++ packages/ng-infrastructure/package.json
@@ -1,11 +1,11 @@
 {
   "name": "@acontplus/ng-infrastructure",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "description": "Acontplus Angular Infrastructure library for managing HTTP interceptors and repositories.",
   "peerDependencies": {
     "@acontplus/core": "^1.0.15",
-    "@acontplus/ng-config": "^1.0.3",
-    "@acontplus/ng-notifications": "^1.0.8",
+    "@acontplus/ng-config": "^1.0.4",
+    "@acontplus/ng-notifications": "^1.0.9",
     "@acontplus/ng-components": "^1.0.13",
     "@angular/common": "^20.3.2",
     "@angular/core": "^20.3.2"
