+++ packages/ng-infrastructure/src/lib/repositories/index.ts
@@ -1,4 +1,3 @@
 export * from './base-http.repository';
 export * from './generic.repository';
 export * from './repository-factory';
-export * from './user.repository';
